fix(sorting_algorithm): minor fixes

+++ src/sorting_algorithm/insertion_sort.c
@@ -6,7 +6,7 @@ void dsInsertionSort(void *arr, unsigned size, size_t elem_size,
 {
 	int i, j;
 
-	for (i = 1; i < size; i++) {
+	for (i = 1; i < (int)size; i++) {
 		for (j = i - 1; j >= 0; j--) {
 			if (compare((char *)arr + (j * elem_size),
 				    (char *)arr + ((j + 1) * elem_size)) > 0) {
